+++ term2048/game.py
@@ -1,7 +1,7 @@
 # -*- coding: UTF-8 -*-
 
 """
-Game logic
+Game logic.
 """
 
 from __future__ import print_function
@@ -12,7 +12,13 @@ import os.path
 import math
 import sys
 
+try:
+    import typing
+except ImportError:
+    pass  # Python 2.6 and 3.4
+
 from colorama import init, Fore, Style
+
 init(autoreset=True)
 
 from term2048 import keypress
@@ -21,31 +27,31 @@ from term2048.board import Board
 
 class Game(object):
     """
-    A 2048 game
+    A 2048 game.
     """
 
-    __dirs = {
-        keypress.UP:      Board.UP,
-        keypress.DOWN:    Board.DOWN,
-        keypress.LEFT:    Board.LEFT,
-        keypress.RIGHT:   Board.RIGHT,
-        keypress.SPACE:   Board.PAUSE,
+    __directions = {
+        keypress.UP: Board.UP,
+        keypress.DOWN: Board.DOWN,
+        keypress.LEFT: Board.LEFT,
+        keypress.RIGHT: Board.RIGHT,
+        keypress.SPACE: Board.PAUSE,
     }
 
     __is_windows = os.name == 'nt'
 
     COLORS = {
-        2:    Fore.GREEN,
-        4:    Fore.BLUE + Style.BRIGHT,
-        8:    Fore.CYAN,
-        16:   Fore.RED,
+        2: Fore.GREEN,
+        4: Fore.BLUE + Style.BRIGHT,
+        8: Fore.CYAN,
+        16: Fore.RED,
         # Don't use MAGENTA directly; it doesn't display well on Windows.
         # see https://github.com/bfontaine/term2048/issues/24
-        32:  Fore.MAGENTA + Style.BRIGHT,
-        64:   Fore.CYAN,
-        128:  Fore.BLUE + Style.BRIGHT,
-        256:  Fore.MAGENTA + Style.BRIGHT,
-        512:  Fore.GREEN,
+        32: Fore.MAGENTA + Style.BRIGHT,
+        64: Fore.CYAN,
+        128: Fore.BLUE + Style.BRIGHT,
+        256: Fore.MAGENTA + Style.BRIGHT,
+        512: Fore.GREEN,
         1024: Fore.RED,
         2048: Fore.YELLOW,
         # just in case people set an higher goal they still have colors
@@ -70,18 +76,20 @@ class Game(object):
 
     def __init__(self, scores_file=SCORES_FILE, colors=None,
                  store_file=STORE_FILE, clear_screen=True,
-                 mode=None, azmode=False, **kws):
+                 mode=None, azmode=False, **kwargs):
         """
         Create a new game.
-            scores_file: file to use for the best score (default
-                         is ~/.term2048.scores)
-            colors: dictionnary with colors to use for each tile
-            store_file: file that stores game session's snapshot
-            mode: color mode. This adjust a few colors and can be 'dark' or
-                  'light'. See the adjustColors functions for more info.
-            other options are passed to the underlying Board object.
+
+        :param scores_file: file to use for the best score (default is ~/.term2048.scores)
+        :param colors: dictionary with colors to use for each tile
+        :param store_file: file that stores game session's snapshot
+        :param clear_screen: if True (the default), clear the screen between each move.
+        :param mode: color mode. This adjusts a few colors and can be 'dark' or 'light'.
+          See the adjustColors functions for more info.
+        :param azmode: if True, use upper-case letters instead of numbers in the grid.
+        :param kwargs: other options are passed to the underlying Board object.
         """
-        self.board = Board(**kws)
+        self.board = Board(**kwargs)
         self.score = 0
         self.scores_file = scores_file
         self.store_file = store_file
@@ -108,48 +116,48 @@ class Game(object):
 
     def loadBestScore(self):
         """
-        load local best score from the default file
+        Load local best score from the default file.
         """
         try:
             with open(self.scores_file, 'r') as f:
                 self.best_score = int(f.readline(), 10)
-        except:
+        except Exception:
             return False
         return True
 
     def saveBestScore(self):
         """
-        save current best score in the default file
+        Save current best score in the default file.
         """
         if self.score > self.best_score:
             self.best_score = self.score
         try:
             with open(self.scores_file, 'w') as f:
                 f.write(str(self.best_score))
-        except:
+        except Exception:
             return False
         return True
 
-    def incScore(self, pts):
+    def incScore(self, score_increment):  # type: (int) -> None
         """
-        update the current score by adding it the specified number of points
+        Update the current score by adding it the specified number of points.
         """
-        self.score += pts
+        self.score += score_increment
         if self.score > self.best_score:
             self.best_score = self.score
 
     def readMove(self):
         """
-        read and return a move to pass to a board
+        Read and return a move to pass to a board.
         """
         k = keypress.getKey()
-        return Game.__dirs.get(k)
+        return Game.__directions.get(k)
 
     def store(self):
         """
-        save the current game session's score and data for further use
+        Save the current game session's score and data for further use.
         """
-        size = self.board.SIZE
+        size = self.board.size
         cells = []
 
         for i in range(size):
@@ -161,23 +169,23 @@ class Game(object):
         try:
             with open(self.store_file, 'w') as f:
                 f.write(score_str)
-        except:
+        except Exception:
             return False
         return True
 
     def restore(self):
         """
-        restore the saved game score and data
+        Restore the saved game score and data.
         """
 
-        size = self.board.SIZE
+        size = self.board.size
 
         try:
             with open(self.store_file, 'r') as f:
                 lines = f.readlines()
                 score_str = lines[0]
                 self.score = int(lines[1])
-        except:
+        except Exception:
             return False
 
         score_str_list = score_str.split(' ')
@@ -216,7 +224,7 @@ class Game(object):
 
     def loop(self):
         """
-        main game loop. returns the final score.
+        Main game loop. returns the final score.
         """
         pause_key = self.board.PAUSE
         margins = {'left': 4, 'top': 4, 'bottom': 4}
@@ -252,51 +260,45 @@ class Game(object):
         print('You won!' if self.board.won() else 'Game Over')
         return self.score
 
-    def getCellStr(self, x, y):  # TODO: refactor regarding issue #11
+    # TODO: refactor regarding issue #11
+    def getCellStr(self, x, y):  # type: (int, int) -> str
         """
-        return a string representation of the cell located at x,y.
+        Return a string representation of the cell located at x,y.
         """
-        c = self.board.getCell(x, y)
+        cell_value = self.board.getCell(x, y)
 
-        if c == 0:
+        if cell_value == 0:
             return '.' if self.__azmode else '  .'
-
         elif self.__azmode:
-            az = {}
-            for i in range(1, int(math.log(self.board.goal(), 2))):
-                az[2 ** i] = chr(i + 96)
-
-            if c not in az:
-                return '?'
-            s = az[c]
-        elif c == 1024:
+            # 2 -> 'a', 4 -> 'b', 8 -> 'c', etc.
+            s = chr(int(math.log(cell_value, 2)) + 96)
+        elif cell_value == 1024:
             s = ' 1k'
-        elif c == 2048:
+        elif cell_value == 2048:
             s = ' 2k'
         else:
-            s = '%3d' % c
+            s = '%3d' % cell_value
 
-        return self.__colors.get(c, Fore.RESET) + s + Style.RESET_ALL
+        return self.__colors.get(cell_value, Fore.RESET) + s + Style.RESET_ALL
 
-    def boardToString(self, margins=None):
+    def boardToString(self, margins=None):  # type: (typing.Optional[typing.Dict[str, int]]) -> str
         """
         return a string representation of the current board.
         """
         if margins is None:
             margins = {}
 
-        b = self.board
-        rg = range(b.size())
-        left = ' '*margins.get('left', 0)
-        s = '\n'.join(
-            [left + ' '.join([self.getCellStr(x, y) for x in rg]) for y in rg])
-        return s
+        rg = list(range(self.board.size))
+        left = ' ' * margins.get('left', 0)
+        return '\n'.join([left + ' '.join([self.getCellStr(x, y)
+                                           for x in rg])
+                          for y in rg])
 
     def __str__(self, margins=None):
         if margins is None:
             margins = {}
         b = self.boardToString(margins=margins)
-        top = '\n'*margins.get('top', 0)
-        bottom = '\n'*margins.get('bottom', 0)
+        top = '\n' * margins.get('top', 0)
+        bottom = '\n' * margins.get('bottom', 0)
         scores = ' \tScore: %5d  Best: %5d\n' % (self.score, self.best_score)
         return top + b.replace('\n', scores, 1) + bottom